+++ MaxMind.MinFraud.UnitTest/Request/TestHelper.cs
@@ -85,7 +85,7 @@ namespace MaxMind.MinFraud.UnitTest.Request
                         ),
                 order:
                     new Order(
-                        amount: (decimal)323.21,
+                        amount: 323.21m,
                         currency: "USD",
                         discountCode: "FIRST",
                         affiliateId: "af12",
@@ -100,13 +100,13 @@ namespace MaxMind.MinFraud.UnitTest.Request
                         category: "pets",
                         itemId: "ad23232",
                         quantity: 2,
-                        price: (decimal) 20.43
+                        price: 20.43m
                         ),
                     new ShoppingCartItem(
                         category: "beauty",
                         itemId: "bst112",
                         quantity: 1,
-                        price: (decimal) 100.00
+                        price: 100.00m
                         )
                 }
                 );
